// backend/src/expenses/expense-category.enum.ts
export enum ExpenseCategory {
  ELECTRICITY = 'ELECTRICITY',
  WATER = 'WATER',
  INTERNET = 'INTERNET',
  RENT = 'RENT',
  MAINTENANCE = 'MAINTENANCE',
  CLEANING = 'CLEANING',
  STATIONERY = 'STATIONERY',
  TRANSPORTATION = 'TRANSPORTATION',
  MARKETING = 'MARKETING',
  EQUIPMENT = 'EQUIPMENT',
  MISC = 'MISC',
}